requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
